+++ fixer-app-backend/src/main/java/org/example/fixerappbackend/service/UsuarioService.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface UsuarioService {
     List<Usuario> findAll();
 
-    Optional<Usuario> findById(Long id);
+    Optional<Usuario> findById(Integer id);
 }
